+++ src/bouillon/git.py
@@ -11,30 +11,39 @@ from __future__ import annotations
 import os
 import subprocess
 
-import bouillon
+from bouillon import run
 
 
 def repository_name() -> str:
     """Get git repository name."""
-    r = bouillon.run(["git", "config", "--get", "remote.origin.url"],
+    r = run(["git", "config", "--get", "remote.origin.url"],
                      stdout=subprocess.PIPE,
                      check=True)
 
-    return str(os.path.split(r.stdout.decode().rstrip())[-1].split(".")[0])
+    return os.path.split(r.stdout.decode().rstrip())[-1].split(".")[0]
 
 
 def current_branch() -> str:
     """Get git current branch."""
-    r = bouillon.run(["git", "rev-parse", "--abbrev-ref", "HEAD"],
+    r = run(["git", "rev-parse", "--abbrev-ref", "HEAD"],
                      stdout=subprocess.PIPE,
                      check=True)
 
-    return str(r.stdout.decode().rstrip())
+    return r.stdout.decode().rstrip()
+
+
+def default_branch() -> str:
+    """Get git default branch."""
+    r = run(["git", "rev-parse", "--abbrev-ref", "origin/HEAD"],
+                     stdout=subprocess.PIPE,
+                     check=True)
+
+    return r.stdout.decode().rstrip().lstrip("origin/")
 
 
 def working_directory_clean() -> bool:
     """Check if the working directory is clean."""
-    r = bouillon.run(["git", "diff", "--quiet", "--exit-code"],
+    r = run(["git", "diff", "--quiet", "--exit-code"],
                      stdout=subprocess.PIPE,
                      check=True)
 
@@ -43,7 +52,7 @@ def working_directory_clean() -> bool:
 
 def commit_id() -> str:
     """Get current git commit id."""
-    r = bouillon.run(["git", "rev-parse", "HEAD"],
+    r = run(["git", "rev-parse", "HEAD"],
                      stdout=subprocess.PIPE,
                      check=True)
 
@@ -52,7 +61,7 @@ def commit_id() -> str:
 
 def tags() -> list[str]:
     """Get list of all git tags."""
-    r = bouillon.run(["git", "tag"],
+    r = run(["git", "tag"],
                      stdout=subprocess.PIPE,
                      check=True)