+++ test/bouillon/test_git.py
@@ -20,6 +20,11 @@ def test_current_branch():
     assert isinstance(bouillon.git.current_branch(), str)
 
 
+def test_default_branch():
+    # assert isinstance(bouillon.git.default_branch(), str)
+    assert bouillon.git.default_branch() == "master"
+
+
 def test_commit_id():
     assert len(bouillon.git.commit_id()) == 40